+++ packages/ebsr/configure/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.10.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@9.9.1...@pie-element/ebsr-configure@9.10.0) (2024-04-22)
+
+
+### Features
+
+* update pie-lib/pie-toolbox PD-2449 ([2a0b40a](https://github.com/pie-framework/pie-elements/commit/2a0b40a826f0338b5eda2e8301f79870be3d0073))
+
+
+
+
+
+## [9.9.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@9.9.0...@pie-element/ebsr-configure@9.9.1) (2024-04-22)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3644, PD-3673 ([5e28ce0](https://github.com/pie-framework/pie-elements/commit/5e28ce0dab4a7ecd93ef7bba45ddb20f768b450b))
+
+
+
+
+
+# [9.9.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@9.8.5...@pie-element/ebsr-configure@9.9.0) (2024-04-18)
+
+
+### Features
+
+* update pie-lib/pie-toolbox SC-30530, SC-30682, PD-3607, PD-3655 ([f08bccf](https://github.com/pie-framework/pie-elements/commit/f08bccf86fcb430e75e410116b000b3affc252c0))
+
+
+
+
+
+## [9.8.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@9.8.4...@pie-element/ebsr-configure@9.8.5) (2024-04-16)
+
+**Note:** Version bump only for package @pie-element/ebsr-configure
+
+
+
+
+
 ## [9.8.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr-configure@9.8.3...@pie-element/ebsr-configure@9.8.4) (2024-04-16)
 
 **Note:** Version bump only for package @pie-element/ebsr-configure