+++ packages/matrix/configure/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.25.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.24.1...@pie-element/matrix-configure@0.25.0) (2024-04-22)
+
+
+### Features
+
+* update pie-lib/pie-toolbox PD-2449 ([2a0b40a](https://github.com/pie-framework/pie-elements/commit/2a0b40a826f0338b5eda2e8301f79870be3d0073))
+
+
+
+
+
+## [0.24.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.24.0...@pie-element/matrix-configure@0.24.1) (2024-04-22)
+
+
+### Bug Fixes
+
+* update pie-lib/pie-toolbox PD-3644, PD-3673 ([5e28ce0](https://github.com/pie-framework/pie-elements/commit/5e28ce0dab4a7ecd93ef7bba45ddb20f768b450b))
+
+
+
+
+
+# [0.24.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.23.1...@pie-element/matrix-configure@0.24.0) (2024-04-18)
+
+
+### Features
+
+* update pie-lib/pie-toolbox SC-30530, SC-30682, PD-3607, PD-3655 ([f08bccf](https://github.com/pie-framework/pie-elements/commit/f08bccf86fcb430e75e410116b000b3affc252c0))
+
+
+
+
+
 ## [0.23.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.23.0...@pie-element/matrix-configure@0.23.1) (2024-04-01)
 
 **Note:** Version bump only for package @pie-element/matrix-configure